requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-crc32c==1.5.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.10.0
multidict==6.0.4
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
pyee==9.1.0
Pygments==2.15.1
PyJWT==2.7.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
streamlit==1.22.0
streamlit-webrtc==0.45.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
twilio==8.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.15.0
